Move syncutil code into sync.py.
